Adding mobile menu close button
The site now uses two custom nav walkers.
One for the desktop nav, and one for the mobile nav.
For some reason, I wasn't able to have them reside in their own file.
The mobile nav class could not be found!
So they are currently both residing in 'nav-walker.php'.