Add custom search results page
Reusing the styling for the mobile nav search form for
the search form on the search page.
This required copying the css for this component into the child theme,
as this css is disabled by the parent theme for screens over
1000px in width.

Created a `search_result.php` template.

Copied across the spyglass pngs from the parent template.

Set the excerpt length to 30 words as the search results page
requires that the excerpt length is 30 words
and not the default length;