Create a news archive page
The 'news' archive page shows a preview for the news custom post type
as well as the report custom post types.
Consequently, the archive.php page is actually the template being rendered
for the news archive, as opposed to something like news.php

Previews will display a thumbnail of the featured image for the news post,
or if it is a report, it will show a thumbnail of the associated PDF.
If there is no available image, it will show a placeholder image.
Images can be specified to be a placeholder image by adding it
to the 'Placeholder' category on the Edit Media page.